Fixed problem with `types` and moving some stuff around
