--- proto/README.md
@@ -1,7 +1,0 @@
-# Protocol buffer definition files (`.proto`) for NEAR on StreamingFast products.
-
-## Usage
-
-Heads on to [Google Protocol Buffer Tutorials](https://developers.google.com/protocol-buffers/docs/tutorials), choose
-your language of choice, then follow instructions in `Compiling your protocol buffers` section about how to
-generate the source files for your language.
