+++ proto/sf/near/type/v1/type.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package sf.near.codec.v1;
+package sf.near.type.v1;
 
-option go_package = "github.com/streamingfast/firehose-near/pb/sf/near/codec/v1;pbcodec";
+option go_package = "github.com/streamingfast/firehose-near/types/pb/sf/near/type/v1;pbnear";
 
 message Block {
   string author = 1;
