+++ types/block.go
@@ -2,14 +2,14 @@ package types
 
 import (
 	"fmt"
+
 	"github.com/golang/protobuf/proto"
 	"github.com/streamingfast/bstream"
-	pbcodec "github.com/streamingfast/firehose-near/pb/sf/near/codec/v1"
 	pbnear "github.com/streamingfast/firehose-near/types/pb/sf/near/type/v1"
 	pbbstream "github.com/streamingfast/pbgo/sf/bstream/v1"
 )
 
-func BlockFromProtoCodec(b *pbcodec.Block) (*bstream.Block, error) {
+func BlockFromProtoCodec(b *pbnear.Block) (*bstream.Block, error) {
 	content, err := proto.Marshal(b)
 	if err != nil {
 		return nil, fmt.Errorf("unable to marshal to binary form: %s", err)
@@ -33,10 +33,10 @@ func BlockFromProtoNear(b *pbnear.Block) (*bstream.Block, error) {
 
 	block := &bstream.Block{
 		Id:             b.ID(),
-		Number:         b.Number(),
+		Number:         b.Num(),
 		PreviousId:     b.PreviousID(),
 		Timestamp:      b.Time(),
-		LibNum:         b.Number() - 1,
+		LibNum:         b.LIBNum(),
 		PayloadKind:    pbbstream.Protocol_NEAR,
 		PayloadVersion: 1,
 	}
