+++ types/go.mod
@@ -4,8 +4,8 @@ go 1.16
 
 require (
 	github.com/golang/protobuf v1.5.2
+	github.com/mr-tron/base58 v1.2.0
 	github.com/streamingfast/bstream v0.0.2-0.20220831142019-0a0b0caa04c3
 	github.com/streamingfast/pbgo v0.0.6-0.20220629184423-cfd0608e0cf4
-	github.com/streamingfast/firehose-near v0.3.0
 	google.golang.org/protobuf v1.27.1
 )
