+++ types/pb/generate.sh
@@ -17,15 +17,17 @@ ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && cd ../.. && pwd )"
 
 # Protobuf definitions
 PROTO=${1:-"$ROOT/../proto"}
-PROTO_ACME=${2:-"$ROOT/proto"}
+PROTO_NEAR=${2:-"$ROOT/proto"}
 
 function main() {
   checks
 
   set -e
+
   cd "$ROOT/types/pb" &> /dev/null
 
-  generate "sf/acme/type/v1/type.proto"
+  generate "sf/near/type/v1/type.proto"
+  generate "sf/near/transform/v1/transform.proto"
 
   echo "generate.sh - `date` - `whoami`" > ./last_generate.txt
   echo "streamingfast/proto revision: `GIT_DIR=$PROTO/.git git rev-parse HEAD`" >> ./last_generate.txt
@@ -42,7 +44,7 @@ function generate() {
     fi
 
     for file in "$@"; do
-      protoc -I$PROTO -I$PROTO_ACME \
+      protoc -I$PROTO -I$PROTO_NEAR \
         --go_out=. --go_opt=paths=source_relative \
         --go-grpc_out=. --go-grpc_opt=paths=source_relative,require_unimplemented_servers=false \
          $base$file
@@ -60,14 +62,14 @@ function checks() {
     echo ""
     echo "To fix your problem, perform those commands:"
     echo ""
-    echo "  go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.25.0"
-    echo "  go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.1.0"
+    echo "  go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.26.0"
+    echo "  go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.2.0"
     echo ""
     echo "If everything is working as expetcted, the command:"
     echo ""
     echo "  protoc-gen-go --version"
     echo ""
-    echo "Should print 'protoc-gen-go v1.25.0' (if it just hangs, you don't have the correct version)"
+    echo "Should print 'protoc-gen-go v1.26.0' (if it just hangs, you don't have the correct version)"
     exit 1
   fi
 }
