+++ types/pb/last_generate.txt
@@ -1,3 +1,3 @@
-generate.sh - Mon Jul 25 16:12:41 EDT 2022 - maoueh
-streamingfast/proto revision: aace594d959fbd3952377cea25c92a92316c35a8
-streamingfast/firehose-near/proto revision: fc92175
+generate.sh - Thu Sep  1 23:15:35 EDT 2022 - maoueh
+streamingfast/proto revision: 45b710949c29447279a2ad27b1c17e2546a70f6b
+streamingfast/firehose-near/proto revision: d5521aa
