+++ types/pb/sf/near/type/v1/type.go
@@ -1,21 +1,36 @@
 package pbnear
 
 import (
+	"encoding/hex"
 	"time"
+
+	"github.com/mr-tron/base58"
 )
 
-func (b *Block) ID() string {
-	return b.Hash
+func (x *Block) ID() string {
+	return x.Header.Hash.AsString()
 }
 
-func (b *Block) Number() uint64 {
-	return b.Height
+func (x *Block) Num() uint64 {
+	return x.Header.Height
 }
 
-func (b *Block) PreviousID() string {
-	return b.PrevHash
+func (x *Block) LIBNum() uint64 {
+	return x.Header.LastFinalBlockHeight
 }
 
-func (b *Block) Time() time.Time {
-	return time.Unix(0, int64(b.Timestamp)).UTC()
+func (x *Block) PreviousID() string {
+	return x.Header.PrevHash.AsString()
+}
+
+func (x *Block) Time() time.Time {
+	return time.Unix(0, int64(x.Header.TimestampNanosec)).UTC()
+}
+
+func (x *CryptoHash) AsString() string {
+	return hex.EncodeToString(x.Bytes)
+}
+
+func (x *CryptoHash) AsBase58String() string {
+	return base58.Encode(x.Bytes)
 }
